Replace anonymous test classes with named fixture classes
