+++ element/qrcode/tests/element_test.php
@@ -33,9 +33,12 @@ use mod_customcert\element\form_element_interface;
 use mod_customcert\element\persistable_element_interface;
 use mod_customcert\element\renderable_element_interface;
 use mod_customcert\element\validatable_element_interface;
-use mod_customcert\element\stylable_element_interface;
-use mod_customcert\service\element_renderer;
+use mod_customcert\tests\fixtures\spy_element_renderer;
 use stdClass;
+
+defined('MOODLE_INTERNAL') || die();
+
+require_once(__DIR__ . '/../../../tests/fixtures/spy_element_renderer.php');
 
 /**
  * Unit tests for the qrcode element.
@@ -182,49 +185,7 @@ final class element_test extends advanced_testcase {
         $el = element::from_record($this->make_record());
 
         // Build a spy renderer that fails the test if render_html() is ever called.
-        $renderer = new class implements element_renderer {
-            /** @var bool */
-            public bool $called = false;
-            /**
-             * Render the element into a PDF context.
-             *
-             * @param \mod_customcert\element\renderable_element_interface $element
-             * @param \pdf $pdf
-             * @param bool $preview
-             * @param \stdClass $user
-             * @return void
-             */
-            public function render_pdf(
-                \mod_customcert\element\renderable_element_interface $element,
-                \pdf $pdf,
-                bool $preview,
-                \stdClass $user
-            ): void {
-                $element->render($pdf, $preview, $user, $this);
-            }
-            /**
-             * Render the element into HTML; records that it was called.
-             *
-             * @param \mod_customcert\element\renderable_element_interface $element
-             * @return string
-             */
-            public function render_html(\mod_customcert\element\renderable_element_interface $element): string {
-                $this->called = true;
-                return '';
-            }
-            /**
-             * Render common content (no-op spy).
-             *
-             * @param stylable_element_interface $element The element. Must also implement layout_element_interface.
-             * @param string $content
-             * @return void
-             */
-            public function render_content(
-                stylable_element_interface $element,
-                string $content
-            ): void {
-            }
-        };
+        $renderer = new spy_element_renderer();
 
         $pdf = $this->getMockBuilder(\pdf::class)->disableOriginalConstructor()->getMock();
         $el->render($pdf, false, new stdClass(), $renderer);
